+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -3,14 +3,9 @@ package techcourse.myblog.web;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import techcourse.myblog.dto.UserDto;
 import techcourse.myblog.dto.UserPublicInfoDto;
 import techcourse.myblog.service.UserService;
-import techcourse.myblog.service.exception.NotFoundUserException;
-import techcourse.myblog.service.exception.SignUpException;
-import techcourse.myblog.service.exception.UserDeleteException;
-import techcourse.myblog.service.exception.UserUpdateException;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -70,26 +65,5 @@ public class UserController {
     private boolean isLoggedInUser(HttpSession httpSession, Long id) {
         UserPublicInfoDto loggedInUser = (UserPublicInfoDto) httpSession.getAttribute(LOGGED_IN_USER);
         return (loggedInUser != null) && loggedInUser.getId().equals(id);
-    }
-
-    @ExceptionHandler({NotFoundUserException.class, UserDeleteException.class})
-    public String handleNotFoundUserException() {
-        return "redirect:/";
-    }
-
-    @ExceptionHandler(SignUpException.class)
-    public String handleSignUpException(Model model, Exception e) {
-        model.addAttribute("errorMessage", e.getMessage());
-        return "sign-up";
-    }
-
-    @ExceptionHandler(UserUpdateException.class)
-    public String handleUpdateUserException(Exception e,
-                                            HttpServletRequest httpServletRequest,
-                                            RedirectAttributes redirectAttributes) {
-        HttpSession httpSession = httpServletRequest.getSession();
-        UserPublicInfoDto user = (UserPublicInfoDto) httpSession.getAttribute(LOGGED_IN_USER);
-        redirectAttributes.addFlashAttribute("errorMessage", e.getMessage());
-        return "redirect:/mypage/" + user.getId() + "/edit";
     }
 }
